requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
pillow
requests
beautifulsoup4
scikit-learn
selenium
